Review feedback: move policy methods to model class
